refactor: extract EpochOverviewHeader component
